+++ common/src/main/java/net/mehvahdjukaar/supplementaries/integration/QuarkCompat.java
@@ -14,8 +14,11 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.structure.Structure;
+import net.minecraft.world.level.saveddata.maps.MapDecoration;
 import net.minecraft.world.phys.AABB;
 import org.jetbrains.annotations.Contract;
+
+import javax.annotation.Nullable;
 
 public class QuarkCompat {
 
@@ -111,7 +114,9 @@ public class QuarkCompat {
     }
 
     @ExpectPlatform
-    public static ItemStack makeAdventurerQuill(ServerLevel level, TagKey<Structure> tag) {
+    public static ItemStack makeAdventurerQuill(ServerLevel serverLevel, TagKey<Structure> destination,
+                                                int radius, boolean skipKnown, int zoom,
+                                                MapDecoration.Type destinationType, @Nullable String name, int color) {
         throw new AssertionError();
     }
 }