+++ fabric/src/main/java/net/mehvahdjukaar/supplementaries/integration/fabric/CompatHandlerClientImpl.java
@@ -2,7 +2,7 @@ package net.mehvahdjukaar.supplementaries.integration.fabric;
 
 public class CompatHandlerClientImpl {
 
-    public static void doSetup() {
+    public static void loaderSpecificSetup() {
     }
 
     public static void init() {