+++ forge/src/main/java/net/mehvahdjukaar/supplementaries/mixins/forge/TreasureMapForEmeraldsMixin.java
@@ -2,12 +2,13 @@ package net.mehvahdjukaar.supplementaries.mixins.forge;
 
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.mehvahdjukaar.supplementaries.integration.CompatHandler;
-import net.mehvahdjukaar.supplementaries.integration.forge.quark.AdventurersQuillItem;
+import net.mehvahdjukaar.supplementaries.integration.QuarkCompat;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.TagKey;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.npc.VillagerTrades;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.trading.MerchantOffer;
@@ -20,9 +21,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(targets = "net.minecraft.world.entity.npc.VillagerTrades.TreasureMapForEmeralds")
+@Mixin(targets = {"net.minecraft.world.entity.npc.VillagerTrades$TreasureMapForEmeralds"})
 public abstract class TreasureMapForEmeraldsMixin {
-
     @Shadow
     @Final
     private TagKey<Structure> destination;
@@ -48,11 +48,14 @@ public abstract class TreasureMapForEmeraldsMixin {
     public void turnToQuill(Entity trader, RandomSource random, CallbackInfoReturnable<MerchantOffer> cir) {
         if (trader.level instanceof ServerLevel serverLevel) {
             if (CompatHandler.QUARK && CommonConfigs.Tweaks.REPLACE_VANILLA_MAPS.get()) {
-                ItemStack map = AdventurersQuillItem.forStructure(serverLevel, this.destination,
+                ItemStack map = QuarkCompat.makeAdventurerQuill(serverLevel, this.destination,
                         100, true, 2, this.destinationType, null, 0);
                 map.setHoverName(Component.translatable(this.displayName));
-
-                cir.setReturnValue(new MerchantOffer(new ItemStack(Items.EMERALD, this.emeraldCost), new ItemStack(Items.COMPASS), map , this.maxUses, this.villagerXp, 0.2F));
+                int uses = CommonConfigs.Tweaks.QUILL_MAX_TRADES.get();
+                int xp = (int) ((this.villagerXp * this.maxUses) / (float) uses);
+                int cost = (int) (this.emeraldCost * CommonConfigs.Tweaks.QUILL_TRADE_PRICE_MULT.get());
+                cir.setReturnValue(new MerchantOffer(new ItemStack(Items.EMERALD, cost),
+                        new ItemStack(Items.COMPASS), map, uses, xp, 0.2F));
             }
         }
     }